Derive listen port from environment variable
Rather than using the fixed port number 8000, check the environment variable PORT, and if it's number-like, use its floor. Otherwise, defer to the default of 8000.

This is in line with the behavior offered by craco and webpack for the frontend, and is a feature the runtime will use to run two backends: main and develop branches.