Partial port of initial data to configuration system
Copied actual data from DBDataUtil to config.js according to types created for purpose (logical types).

Logical types are called that since they map to real table types in a non-trivial way.

This branch is not finished, and this commit also includes other changes otherwise uncommitted in the interest of finalisation.